config, catalog: implement viewing of books, upd. Bootstrap
Implement simple viewing of book and a catalog.
Implement REST-controller for getting a file.
Implement minimal configuration of service that works with files.
Add some data to the database and to the file bucket directory.
Add main fragments of the page to a separate html file.
Implement mappers, extractors, statement creator, entities and DTOs for
viewing a catalog and etc.