config, filter, interceptor, cart: implement cart page and REST api for cart.
Implement cart functionality for the GUEST only.
Add an interceptor for setting a context path variable in each page.
Add a demo-filter for setting a userdetails after first access.
Add a filter for setting a cart to the session (only for GUEST now)
Add a validators.
Implement an error handlers for REST controllers.
Add a Role entity for future user.